Significant improvements to find_package() wrapper.
If the package we've been asked to find is IN_TREE, then set up the
environment usually set up by find_package() proper, and then execute
find_dependency commands for all of the package's dependencies. Fail if
we're using a version of CMake too new to be supported. Take the result
of the find_dependency() attempts into account in the values of
${PKG}_FOUND and ${PKG_UC}_FOUND.